+++ src/test/scala/zio/pravega/CommonTestSettings.scala
@@ -6,8 +6,9 @@ import io.pravega.client.stream.Serializer
 import java.nio.ByteBuffer
 
 import scala.language.postfixOps
+import model.Person
 
-object CommonSettings {
+object CommonTestSettings {
 
   val intSerializer = new Serializer[Int] {
     override def serialize(value: Int): ByteBuffer = {
@@ -47,4 +48,27 @@ object CommonSettings {
   )
     .build()
 
+  val personSerializer = new Serializer[Person] {
+
+    override def serialize(person: Person): ByteBuffer =
+      ByteBuffer.wrap(person.toByteArray)
+
+    override def deserialize(buffer: ByteBuffer): Person =
+      Person.parseFrom(buffer.array())
+
+  }
+
+  val personReaderSettings =
+    ReaderSettingsBuilder()
+      .withTimeout(2 seconds)
+      .withSerializer(personSerializer)
+
+  val personStremWritterSettings =
+    WriterSettingsBuilder()
+      .withSerializer(personSerializer)
+
+  val personStremWritterSettingsWithKey =
+    WriterSettingsBuilder[Person]()
+      .withKeyExtractor(_.key)
+      .withSerializer(personSerializer)
 }
