+++ src/test/scala/zio/pravega/EventStreamSpec.scala
@@ -8,6 +8,7 @@ import zio.test.Assertion._
 
 object EventStreamSpec extends SharedPravegaContainerSpec("event-streaming") {
 
+  import CommonTestSettings._
   override def spec: Spec[Environment with TestEnvironment with Scope, Any] =
     scopedSuite(
       test("EventStream support timeouts")(
